fix(devex): tree view cleanup
