+++ frontend/src/layout/navigation-3000/components/ProjectTree/types.ts
@@ -1,0 +1,12 @@
+import { FileSystemEntry } from '~/queries/schema/schema-general'
+
+export interface ProjectTreeAction {
+    type: 'move' | 'move-create' | 'create' | 'delete'
+    item: FileSystemEntry
+    path: string
+    newPath?: string
+}
+
+export interface FileSystemImport extends FileSystemEntry {
+    icon?: JSX.Element
+}
