+++ frontend/src/queries/schema.json
@@ -7363,7 +7363,7 @@
                     "description": "Type of object, used for icon, e.g. feature_flag, insight, etc"
                 }
             },
-            "required": ["id", "path"],
+            "required": ["path"],
             "type": "object"
         },
         "FileSystemType": {
@@ -7380,7 +7380,8 @@
                 "destination",
                 "site_app",
                 "transformation",
-                "folder"
+                "folder",
+                "aichat"
             ],
             "type": "string"
         },
